Add an API to get committed memory usage for local shader compile workers
To better track the total memory usage for shader compilation during cooking, this API gives an estimate of the total committed memory used by local SCWs. This is only an estimate, as you generally cannot sum up the committed memory across processes and assume that this is the total committed memory of the group of processes. But it is the best estimate available.

#rb Laura.Hermanns

[CL 31119959 by sebastian schoner in 5.4 branch]