[RemoteControl] Alternative API to prevent remote control presets from renewing their Guids on duplicate.
The Motion Design Rundown editor is loading and duplicating level assets containing embedded remote control presets. This is done as part of the Rundown editor's work flow in order to have a copy of the RCP that will be used to run the controllers and capture the resulting values. We want the duplicated RCP to keep it's guids the same as the original.

The original method for doing this was to use EDuplicateMode::PIE. However, since there are may be no PIE instances running, this method hits some check failures. One such failure in RemapGuid (LazyObjectPtr.cpp) seems to indicate EDuplicateMode::PIE can't be used if there are no PIE instances running.

To resolve this issue for the moment, we will add an API to explicitly disable the guid renewal.

#rb Juan.Portillo

[CL 31144285 by jean perrier in ue5-main branch]