Add a CVar for propagating local SCW OOMs
When we run in a docker container with a memory limit, every out-of-memory condition should be reported as a total failure. Right now, when a local SCW goes OOM we do not report it as a failure, meaning that our reporting for live failures is inaccurate and under-reports memory issues. For future work, we could consider more explicitly limiting the amount of memory SCWs can use and be more intelligent about handling this, but for the moment this is a worthwhile improvement to correctly assess the scope of the problem.

#rb Laura.Hermanns

[CL 31173854 by sebastian schoner in 5.4 branch]