Report Metis memory errors as global OOMs
When memory is limited, the GraphPartitioner in Nanite's building code can run out memory from within the METIS library. This is currently not properly raised as an OutOfMemory error in our memory subsystem, which means that we are missing those in our error reporting for memory issues. We are already aborting when this happens, so there is no behavior change besides that the error is now properly picked up by our reporting systems.


#rb Brian.Karis

[CL 31207367 by sebastian schoner in 5.4 branch]