Add a CVar for Windows to estimate total memory usage in a container
We are currently using an artificial limit on memory usage when running in a container. This limiting mechanism only accounts for the current process, which might miss gigabytes of memory coming from shader compilation or anything else we run in the container: .NET scripts, crash reporting etc. This error leads us to overestimate available memory, so we don't trigger GC or hold off on starting memory intensive processes.

On Windows, containers are implemented using job objects, which are essentially nestable groups of processes. Job objects allow you to set a limit for total committed memory. Because life is cruel, the API does not allow you to query how much memory you currently have committed. We instead estimate it by going over all processes in the job object and sum up their pagefile usage. This estimate could be both too high (shared memory) or too low (kernel objects associated with the process may not be counted), but empirically it is much more accurate than the current version. In my testing, I have seen the first green runs for projects that have been going OOM for months.

Performance-wise, this change is terrible. I have measured that on a 20min cook we spend about 5s just estimating memory usage. This is bad, but ultimately still preferable to failing a cook because we did not realize that we should trigger a GC, for example. In that case we likely re-run the cook multiple times, wasting significantly more than 5s.

(Note that our memory stats use the term "virtual memory" to refer to "committed memory" on Windows.)



#rb charles.bloom

[CL 31207462 by sebastian schoner in ue5-main branch]