Path Tracing: Fix blurry textures
The derivative calculations in the path tracing material code doesn't appear to be correct in all cases. We have seen some reports of blurry textures that go away when this is disabled. Since the derivatives are mainly used by Substrate glints for now, only enable them in this case.

Also do not hardcode SUBSTRATE_COMPLEXSPECIALPATH since it should only be defined for materials that need it.

[FYI] Charles.deRousiers

#changelist validated
#virtualized

[CL 31217801 by chris kulla in ue5-main branch]