Cameras: rewrite the core camera system architecture using evaluator objects built from the data nodes
The previous approach, which was about data nodes running themselves, and being cloned in the runtime to instantiate them, didn't yield the expected performance levels. It ran well in 95% of cases (under 0.02ms for a representative node tree), but had unexpected spikes of more than 0.5ms in a minority of cases. This new approach yields more consistent performance (max time was around 0.13ms, median time was the same), with many more opportunities for optimization, but at the cost of more boilerplate code to implement new camera nodes.

Camera node evaluators are built by the nodes, and are responsible for running the corresponding logic. Evaluators are stored in chunks of contiguous memory owned by either the camera system itself, or by the blend stack in which they belong. There are opportunities later to pre-compute the space needed for a given camera node hierarchy, so that we can guarantee all those evaluators are stored in a single allocation chunk.

Camera directors also run with a similar design, but their evaluators are simply stored as a shared pointer, since there's only a couple of directors active at one point.

This change also introduces a simple RTTI framework, since we need to safely inspect or downcast evaluators in some cases.

The live-edit manager interface is made empty for now since we don't need to forward property changes to instantiated objects anymore (evaluators for two instances of a camera mode share the same data nodes), but we keep it because we will later need to handle hot-reloading a node hierarchy by re-building its evaluator tree.

#jira UE-205641

[CL 31223336 by ludovic chabant in ue5-main branch]