Update the comments on bIgnoreAllPressedKeysUntilRelease to be clear and provide some examples.
Ensure that we set bShouldBeIgnored AFTER we retain the state of old triggers and modifiers. Otherwise we would never actually update the state of the mapping's bShouldBeIgnored flag and it would always be false. I believe that this issue was introduced with 24566913. Once we got the right behavior for that comparison operator, then the mapping would always be converted instead of the old behavior where it would not. This resulted in it overriding the bShouldBeIgnored flag to whatever the previous state was.

#jira UE-196939
#rb dan.oconnor


#changelist validated
#virtualized

[CL 31224755 by ben hoffman in 5.4 branch]