[MotionDesign] Playback Service Messages: Removing comments about the idea of supporting a playback graph directly on the server side.
The way the playback graph would be implemented on the playack server side was either that it would directly run the playback graph asset, or it would create a transient playback graph that wraps the playable.
The requirements for the Rundown have evolved so that the second approach was heavily favored and the first approach was never used. It can therefore be considered not supported. And thus why indications to the contrary in the comments are removed for now.

#rnx

[CL 31227439 by jean perrier in 5.4 branch]