Fix tile artifact on hair groom.
Primary visibility memory allocation is based on groom CPU bound. When using long hair with physic, the CPU bound might not be conservative enough, resulting into too low hair visibility memory allocation. This CL increase the CPU bound size for this allocation proportionaly to the longest curve of a groom.

#rb charles.derousiers

[CL 31287708 by charles derousiers in ue5-main branch]