Fix incorrect rendering in D3D11 and 12 when losing focus in exclusive fullscreen mode.
Slate switches to Windowed Fullscreen when we lose focus, but doesn't explicitly resize the window, so we never call SetWindowPos or ResizeTarget. If the fullscreen swapchain had a lower resolution than the monitor's native resolution, or if the swapchain had been windowed before, we were stretching the backbuffer to that smaller resolution inside the borderless window, instead of covering all of it. ResizeBuffers is not sufficient because it only affects the buffers, not the window. We need to also call ResizeTarget explicitly when dropping out of exclusive fullscreen, to make sure the window is correctly set up.

Vulkan doesn't have this problem because it doesn't support exclusive fullscreen.

#jira UE-203916
#rnx
#rb benjamin.rouveyrol

[CL 31290487 by mihnea balta in ue5-main branch]