Path Tracing: Avoid path tracer invalidation when a camera is selected in the editor
This was creating an editor widget, which was in turn setting some colors on a dynamic material instance which led to the path traced output being invalidated. Simply exclude any materials with bUsedWithEditorCompositing set from invalidating the path tracer. The assumption is that this flag will only be set for specific engine materials used in the UI.

#jira UE-206160

[CL 31345047 by chris kulla in 5.4 branch]