Virtual Scouting - fix for manual focus getting set to 1cm when picking up a camera with the viewfinder tool.
Fixed focus line trace logic to ignore VScout lights, drawings and bookmarks so geo created by users in a session does not block the focus line trace. This is a temporary measure until we can investigate having a custom trace channel for projects using VScout.

Fixed logic around rebuilding the monitor. Viewfinder UMG widget is not destroyed and recreated each time the monitor geo is rebuilt. Widget is only created once, when the tool is first initialized.

Fix in VScout button classes for findparentbywidgettype == null when not parented to the XRCreativePalette
#jira UE-205595, UE-205591
[FYI] tommy.lahitte
#rb richard.graham

#virtualized

[CL 31351544 by richard graham in 5.4 branch]