Color Picker Current Color Theme textbox shows double text
This was happening because there was a hack in SColorPicker.cpp to show the placeholder text for the SThemeColorBlocksBar, which was broken. It has since been fixed, so I removed the hack to fix the placeholder text. At this point the placeholder text was misaligned, so I fixed the padding for it in SColorThemes.cpp. Befpre and after pic is attached to the Jira.


#jira UE-203428
#rb JeanMichel.Dignard

[CL 31354947 by karen jirak in ue5-main branch]