LiveLink: Added BP function to query subject state (invalid, disconnected, unresponsive).
This should mimic the state indicated in the live link panel for a given subject. The function is blueprint callable, and is called GetLiveLinkSubjectState. The value returned is an enum, as follows:

enum class ELiveLinkSubjectState : uint8
{
  Connected,
  Unresponsive,
  Disconnected,
  InvalidOrDisabled,
  Unknown
};

It is similar to ETimedDataInputState but can report more states suitable to a higher level query.

#jira UE-206228
#rb alejandro.arango

[CL 31363829 by Alejandro Arango in ue5-main branch]