Reduce vertical space consumed by title bar
The intent of this change is twofold. First, it fixes a bug where the min/max/close buttons overlapped the row of tabs in windows with a main menu. Second, it reduces the vertical space consumed by our windows in the area around the title bar, main menu, and tabs.

* Reduced height of SVG images used for the min/max/restore/close window buttons from 34 pixels to 23 pixels. This avoids an overlap of the min/max/close buttons and the row of tabs in windows with a main menu.
* Rebalanced padding around the Unreal icon in windows with or without a main menu. There is now at most 1 pixel more spacing below the logo than above where before it could be 3 to 8 pixels more below than above.
* Reduced padding between tabs and main menu.

Note that these changes are all within #if !PLATFORM_MAC, so the Mac should not be affected.

#jira UE-204349
#rb brooke.hubert

[CL 31377272 by sebastian arleryd in 5.4 branch]