GpuProfiler: Support gpu trace events named with the <basename>_N pattern, where N is some integer.
Before this fix, Gpu trace events with said naming pattern would all appear with the same name, which would correspond to the first one. e.g. VP_0, VP_1, VP_2 would all appear as VP_0 in Unreal Insights.

This is because the event is being sent as a uint32 corresponding to the ComparisonIndex of FName, which is the same for the above naming pattern - the N ends up encoded in the Number of the FName, corresponding to the instance number.

The fix in this CL forces a space at the end of the FName when the instance number is greater than 0.

Upgrading the GpuProfiler to include the FName instance number was also possible without breaking backwards compatibility (open old traces with new insights) but adds 4 bytes per event to handle the FName instance number, so went for the simpler solution.

#rb alejandro.arango
[FYI] jason.hoerner

[CL 31461235 by alejandro arango in 5.4 branch]