PR #11156 (modified): [GAS] Fixed bugs caused by gameplay cue getting removed after OnRemove is sent to the cue
Note:  This will change behavior slightly.  Previously, removing a GameplayCue would cause a single EGameplayTagEventType::AnyCountChange callback with its Tag and Count, even though all equivalent Cues would receive an OnRemove.  So previously:  15 GC Tags would be a single callback and now it will be 15 callbacks.  This will change behavior if you're using something like AbilityTask_WaitGameplayTagCountChanged on the GameplayCue Tag (that callback would fire 15x now, and just once previously).  I believe this simplification of the code (along with such an unlikely edge case) warrants the change in behavior.

The CVar to toggle this functionality is GameplayCue.Fix.UseEqualTagCountAndRemovalCallbacks (it is now defaulted to true)

#jira UE-201203
#rb oscar.sweatman, Todd.Eckert

[CL 31487623 by erebel55 in ue5-main branch]