Sequencer: Fixed hierarchical easing channels being destroyed when still in use
The issue here was new entities being linked from inside an existing subsequence that had hierarchical blending enabled (which still only really happens with Day Sequence). In this sitation the number of hierarchical easing providers has not changed, so bChannelsHaveBeenInvalidated is false, but we were previously always destroying all the channels for blend targets inside FinalizeBlendTargets and only re-constructing them from the { NeedsLink | HierarchicalBlendTarget } entities. Clearly this results in pre-existing channels that relate to blend targets being destroyed if they did not happen to have a { NeedsLink | HierarchicalBlendTarget } entity.

There are two options to fix this: firstly we could always re-create and re-assign _all_ easing channels that relate to hierarchical blend targets, or we could ensure that we don't destroy channels that relate to blend targets in this situation. I opted for the latter since the first option is way more expensive.


#rb Max.Chen

[CL 31497814 by andrew rodham in ue5-main branch]