Check for read only channels when enabling channels
Some channels are marked read-only (e.g. memory tracing channels) and cannot be toggled while the application is running. However there was no feedback when users tried to enable those channels using console commands for example or using the public API. With this change a logging channel argument is added to several internal methods in order to output information correctly, to console feedback when needed.

#rb ionut.matasaru
#jira UE-149917

[CL 31514835 by Johan Berg in 5.4 branch]