MoviePipeline: Fixed issue that caused graphs to have their dirty flag set as soon as they are opened.
This was fixed by making the editor representation of the graph non-transient and saved as part of the runtime graph. Before, the editor representation was re-created when the runtime graph was loaded, thus dirtying the graph as soon as it was shown in the editor.

Most of the changes here, aside from making the editor representation non-transient, all revolve around making sure delegates get registered upon *initial* object creation (ie, in a new, non-loaded graph) AND when the object is *loaded* (ie, from an existing graph). This was not a concern before because the graph was always re-created upon load.

#jira UE-202708
#rb Matt.Hoffman
#rnx

[CL 31541686 by chad williams in 5.4 branch]