UE-143412 - Using Collapse to Function inside of a thread safe graph does not mark the resulting function as thread safe, causing a compile error until it is checked
Follow-up fix for threadsafe graphs. The previous fix in CL 30950233 failed to account for native functions that were marked as threadsafe. In this case, we need to check the metadata on the corresponding UFunction instead.

We already have a utilty function that handles both cases, which can be used here: HasFunctionBlueprintThreadSafeMetaData.

#jira UE-143412
#rb dan.oconnor

[CL 31555546 by dave jones2 in ue5-main branch]