Horde: Add firewall exception to Horde server in installer
Flashes a command-line window for a brief second during the install when this command is run. I found no immediate easy way to prevent this behavior.
Wix does support quiet exection through some extension but was unable to get that working. Might be worth a revisit if we found the flashing window to be too intrusive.

#skipci

[CL 31677492 by carl bystrom in ue5-main branch]