Added alert column support to TEDS.
It's now possible to add alerts to any row in TEDS. If the row is used by the Outliner, an icon will show up indicating there are alerts. If the row has a parent column as well, a child alert will be added to the parent row automatically to allow keeping track of the number of children that have alerts. Currently alerts can be either warnings or errors. Console commands were added to support testing, which allow adding/removing alerts to selected rows or in bulk.

#jira UE-206167
#rb logan.buchy

[CL 31694101 by ronald koppers in ue5-main branch]