Cameras: add camera variable assets, and camera parameters that can be driven by these variables
Camera variable assets are ways for users to safely define the name and type of a camera variable, which can then be placed in the variable table. This system will be usable as a "blackboard" of sorts, where gameplay values can be pushed into the camera system and used there to drive camera behaviors. This can be done with camera parameters, which allow camera nodes to accept both fixed values and references to variable assets.

#jira UE-207729

[CL 31708945 by ludovic chabant in ue5-main branch]