Cameras: rename "camera mode" to "camera rig"
While "camera mode" is a somewhat standard term for a specific sort of gameplay camera, it's actually too specific. The term "camera rig" is more generic and can describe other hierarchies of nodes that don't have anything to do with a given gameplay camera behavior, such as those nodes that run before or after the blend stack, or those that are re-used as "prefabs" of sorts. A "camera mode" is therefore only a top-level camera rig that happens to run in the main blend stack.

Of course, another reason for this change is to avoid the easy confusion between "camera mode" and "camera node".

[CL 31715317 by ludovic chabant in ue5-main branch]