Sequencer: Numeric edits to key values now only occur on Enter and not when the focus changes.
This removes a hack that temporarily clears focus before changing time and then restoring focus. The reason that was necessary when key value changes were commited on focus change, the commit would happen after the time changed, which meant that the key would change on the time you're changing to, rather than the current time that you're trying affect. Regardless, this behavior is all removed now.

Tested the original jira bug and it no longer occurs with these changes.

#jira UE-71646
#jira UE-206986
#rb Andrew.Rodham
#lockdown laurent.delayen

[CL 31734641 by max chen in 5.4 branch]