Fix scene query crash when rewind/resim is in use
During rewind/resim we run the solver again to resimulate the required frames. This was invoking the game thread only callbacks on a worker thread. This led to a game thread only flag being set and queries from the game thread being led to believe they were supposed to query against physics thread data.
For now this change disables the ability for the resim step to call game thread callbacks and added some extra safety to ensure we never flip the frozen GT flag on a worker thread in future.

#rb cedric.caillaud, markus.boberg, tom.waterson

[CL 31761915 by benn gallagher in 5.4 branch]