Cameras: change how evaluator trees are built, and add allocation info on camera rigs
Evaluator trees are now built in 2 phases: the first phase actually constructs them, and the second phase initializes them. This separates the two different concerns, but also makes it easier to build a temporary (non-functional) evaluator tree for the purposes of checking how much memory we need to allocate the "real" one. This is what the new allocation builder class does to set the allocation on a camera rig.

In addition to that, this new allocation builder class also traverses the node tree to gather all used camera variables and figure out the optimal variable table layout.

Both the evaluator allocation info and variable table layout info are saved on the camera rig when it is built.

[CL 31792294 by ludovic chabant in ue5-main branch]